% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{my hello world function}
\usage{
hello(x)
}
\arguments{
\item{x}{the name of the person to say hi to}
}
\value{
the output form \code{\link{print}}
}
\description{
my hello world function
}
\examples{
hello("Jhon")
\dontrun{
hello("Steve")
}
}
