% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\docType{data}
\name{pos.csv}
\alias{pos.csv}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\references{

}
\author{
Ahmed AlRashid \email{ahmed.alrashid.04@gmail.com}
}
\keyword{data}
