% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_read.R
\name{simple_read}
\alias{simple_read}
\title{Read CSV file}
\usage{
simple_read(path)
}
\arguments{
\item{path}{path to filename}
}
\value{
\code{dataframe}
}
\description{
Read CSV file
}
\examples{
csv = system.file("extdata", "pos.csv", package = "pospkg")
simple_read(csv)
}
